Neuronment v0.4.3.1 - Reproducing GA5.3 Results - Added V1_D001_weight_factor variable - Modified V1Neurons naming to fit GA5.3 sorting - Edited resources file names to match template - Edited V1V1_Links with the GA5.3 training results for Exp001 - Fully verified V1 external currents calculations - Fully verified V1MT weights calculations - Deep verification of V1V1 weights calculations. Small discrepancies still pending

+++ Neuronment_v1.0/Sources/SimulationManager.cpp
@@ -218,7 +218,7 @@ bool SimulationManager::SetV1V1ConnectionLinks()
     }
     for (int i = 0; i < V1_Neurons.size(); i++) {
       for (int j = 0; j < V1_Neurons.size(); j++) {
-        V1_Neurons[i].SetLinkingWeight(V1_Neurons[j].GetName(), V1_Neurons[i].GetLinkingWeight(V1_Neurons[j].GetName()) / V1_Neurons[i].ConnectionCount(Neuron_V1));
+        V1_Neurons[i].SetLinkingWeight(V1_Neurons[j].GetName(), V1_Neurons[i].GetLinkingWeight(V1_Neurons[j].GetName()) / (double) V1_Neurons[i].ConnectionCount(Neuron_V1));
         Log.Message("WN-006: V1V1 LinkWeight count correction");
       }
     }
